refactor: 파일 평탄화 작업
