--- features/profile/api/queries.ts
@@ -1,14 +1,0 @@
-import { useQuery } from '@tanstack/react-query';
-
-import { fetchProfile } from './fetchers';
-
-export const profileKeys = {
-  all: ['profile'] as const,
-  detail: () => [...profileKeys.all, 'detail'] as const,
-};
-
-export const useProfile = () =>
-  useQuery({
-    queryKey: profileKeys.detail(),
-    queryFn: fetchProfile,
-  });
